Use autoroute for new text entry (and back)
