+++ lotti/lib/routes/journal_routes.dart
@@ -1,4 +1,5 @@
 import 'package:auto_route/auto_route.dart';
+import 'package:lotti/pages/create/create_text_entry_page.dart';
 import 'package:lotti/pages/journal/entry_details_page.dart';
 import 'package:lotti/pages/journal_page.dart';
 
@@ -14,6 +15,10 @@ const AutoRoute journalRoutes = AutoRoute(
     AutoRoute(
       path: ':entryId',
       page: EntryDetailPage,
+    ),
+    AutoRoute(
+      path: 'create/:linkedId',
+      page: CreateTextEntryPage,
     ),
   ],
 );
