+++ lotti/lib/widgets/create/add_actions.dart
@@ -1,5 +1,6 @@
 import 'dart:io';
 
+import 'package:auto_route/auto_route.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:lotti/blocs/audio/recorder_cubit.dart';
@@ -7,7 +8,6 @@ import 'package:lotti/classes/journal_entities.dart';
 import 'package:lotti/get_it.dart';
 import 'package:lotti/logic/image_import.dart';
 import 'package:lotti/logic/persistence_logic.dart';
-import 'package:lotti/pages/add/editor_page.dart';
 import 'package:lotti/pages/add/new_measurement_page.dart';
 import 'package:lotti/pages/add/new_task_page.dart';
 import 'package:lotti/pages/add/survey_page.dart';
@@ -135,15 +135,8 @@ class _RadialAddActionButtonsState extends State<RadialAddActionButtons> {
         ),
         backgroundColor: AppColors.actionColor,
         onPressed: () {
-          Navigator.of(context).push(
-            MaterialPageRoute(
-              builder: (BuildContext context) {
-                return EditorPage(
-                  linked: widget.linked,
-                );
-              },
-            ),
-          );
+          String? linkedId = widget.linked?.meta.id;
+          context.router.pushNamed('/journal/create/$linkedId');
         },
       ),
     );
